backend/requirements.txt
crewai[tools]>=0.193.2,<1.0.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
